requirements.txt
pandas==2.2.2
Pillow==10.4.0
streamlit==1.35.0
